refactor : 파일 정리, 오타수정, 라인정리
