+++ src/main/java/baseball/domain/GameStatus.java
@@ -1,4 +1,4 @@
-package baseball.controller;
+package baseball.domain;
 
 public enum GameStatus {
