+++ src/test/java/baseball/controller/ConsoleViewTest.java
@@ -4,7 +4,7 @@ import jdk.jfr.Description;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
-import utils.Validator;
+import baseball.util.Validator;
 
 class ConsoleViewTest {
